requirements.txt
Flask==2.3.3
cloudinary==1.36.0
python-dotenv==1.0.0
Flask-Cors==4.0.0